+++ src-tauri/src/input/teacher.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct Teacher {
-    pub id: u64,
-    pub index: u64,
+    pub id: usize,
+    pub index: usize,
     pub name: String,
 }
